+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## Patch fixes
+Version: 1.1.0
+Date: 2 Nov, 2025
+
+Updates:
+* Fixed class name consistency: Updated "Initializer" to "MCPInitializer" throughout documentation
+* Removed incorrect "nextCursor" fields from tools/list and prompts/list response examples
+
 ## Initial Documentation Release
 Version: 1.0.0
 Date: 2 Nov, 2025